refactor(TAS-49): userMenu refactoring
