+++ src/modules/headerMenu/types/MenuItem.ts
@@ -1,4 +1,4 @@
-export type MenuItems = {
+export type MenuItem = {
   name: string,
   path: string,
 };
